TODO List updates

+++ examples/todo_list/lib/todo_list/sidebar.rb
@@ -3,6 +3,7 @@ module TodoList
   class Sidebar < Cura::Component::Pack
     
     attr_reader :create_list_textbox
+    attr_reader :listbox
     
     def initialize(attributes={})
       attributes = { fill: true, padding: { top: 1, bottom: 1 } }.merge(attributes)
@@ -16,7 +17,7 @@ module TodoList
       @create_list_textbox.on_event(:key_down, self) { |event, sidebar| sidebar.create_list if event.name == :enter }
       create_form_pack.add_child(@create_list_textbox)
       
-      @create_list_button = Cura::Component::Button.new( width: (width/3).floor, text: 'Create', padding: { left: 1, right: 1 } )
+      @create_list_button = Cura::Component::Button.new( width: (width/3).floor, text: 'Create', alignment: { horizontal: :center } )
       @create_list_button.on_event(:click, self) { |event, sidebar| sidebar.create_list }
       create_form_pack.add_child(@create_list_button)
       
@@ -25,7 +26,7 @@ module TodoList
       
       @listbox = Cura::Component::Listbox.new( width: @width )
       @listbox.on_event(:key_down, self) do |event, sidebar|
-        if event.control? && event.name == :D && !selected_object.nil?
+        if event.target == self && event.control? && event.name == :D && !selected_object.nil?
           selected_object.destroy
 
           previous_selected_index = @selected_index
@@ -38,6 +39,8 @@ module TodoList
           selected_child.focus
         end
       end
+      
+      @listbox.on_event(:selected) { |event| application.list_items.list = selected_object }
       
       add_child(@listbox)
